DIGEQ-159 refactoring validators into seperate class files
